requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
joblib==0.17.0
numpy==1.19.3
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.54.0
urllib3==1.26.2
ipython==7.19.0